Fixed Spawn Point Issue, Colliders changed to 3D instead of 2D
