+++ Assets/ArtAssets/SFX/Particles/Unity samples/particlePack_samples.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a0ec4f01b227abb4db4e881a314960be
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
