+++ Assets/Prefabs/craft_speederB Variant.prefab
@@ -12,6 +12,18 @@ PrefabInstance:
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 8024c44a41c636a4db788f83570d823a, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 8024c44a41c636a4db788f83570d823a, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 8024c44a41c636a4db788f83570d823a, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 8024c44a41c636a4db788f83570d823a, type: 3}
       propertyPath: m_LocalPosition.x
       value: 3.56
       objectReference: {fileID: 0}
@@ -55,6 +67,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: craft_speederB Variant
       objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 8024c44a41c636a4db788f83570d823a, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 8024c44a41c636a4db788f83570d823a, type: 3}
 --- !u!1 &4509982482327964147 stripped
@@ -62,32 +78,35 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 8024c44a41c636a4db788f83570d823a, type: 3}
   m_PrefabInstance: {fileID: 3627584619507463330}
   m_PrefabAsset: {fileID: 0}
---- !u!61 &-3353137909111853579
-BoxCollider2D:
+--- !u!54 &-4421877792245711252
+Rigidbody:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4509982482327964147}
-  m_Enabled: 1
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 1
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0.29999998}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0, y: 0}
-    oldSize: {x: 0, y: 0}
-    newSize: {x: 0, y: 0}
-    adaptiveTilingThreshold: 0
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 2.0000002, y: 0.59999996}
-  m_EdgeRadius: 0
+  m_Mass: 1
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_UseGravity: 0
+  m_IsKinematic: 0
+  m_Interpolate: 0
+  m_Constraints: 4
+  m_CollisionDetection: 0
+--- !u!65 &5884364715491259545
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4509982482327964147}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Size: {x: 2.0000002, y: 0.59999996, z: 2.0256703}
+  m_Center: {x: 0, y: 0.29999998, z: 0}
 --- !u!114 &-6902075080154479171
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -121,4 +140,3 @@ MonoBehaviour:
   maxSpeed: 2
   damage: 1
   health: 5
-  p: {fileID: 0}
