+++ Assets/Scripts/MainGame/EnemyScripts/Enemy_Basci.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Enemy_Heavy : MonoBehaviour
+public class Enemy_Basci : MonoBehaviour
 {
     private float speed;
     public float minSpeed;
@@ -10,26 +10,13 @@ public class Enemy_Heavy : MonoBehaviour
     public int damage;
     public int health;
 
-    //public Player p;
-
-
-
-
     // Start is called before the first frame update
-
-    //enemy need access to player to call the take damage method
     void Start()
-    {
-
-        //p = GameObject.Find("Player");
-        //p = GameObject.FindGameObjectWithTag("Player").GetComponent<Player>();
-        this.transform.rotation = Quaternion.Euler(0f, 180f,0f);
+    {                                                       //2f
+        this.transform.rotation = Quaternion.Euler(0f, 180f, 0f);
         speed = UnityEngine.Random.Range(minSpeed, maxSpeed);
+        Physics.IgnoreLayerCollision(7,7);
     }
-
-
-
-
 
     // Update is called once per frame
     void Update()
@@ -38,23 +25,17 @@ public class Enemy_Heavy : MonoBehaviour
     }
 
     void FixedUpdate()
-    {
-
-        //Debug.Log(p);
-
-        //-1f                 
+    {                                       //-1f                                      , 0.1f)
         this.transform.Translate(new Vector3(1f, UnityEngine.Random.Range(-0.1f, 0.1f),0f) * speed * Time.deltaTime);
-        
-        //
 
         /*if (this.gameObject.transform.position.x <= 0)
         {
             Destroy(this.gameObject);
-            p.takeDamage(damage);
-        }*/
+        }
+*/
     }
 
-    private void OnTriggerEnter2D(Collider2D collision)
+    private void OnTriggerEnter(Collider collision)
     {
         if (collision.tag == "Player")
         {
