--- Assets/Scripts/MainGame/EnemyScripts/SpawnPoint.cs
@@ -1,8 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class SpawnPoint : MonoBehaviour
-{
-    public Vector3 Destination;
-}
