--- Assets/Scripts/MainGame/EnemyScripts/SpawnPoint.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c90a2fcf42d66c448527b64abaf623b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
