+++ Assets/Scripts/MainGame/SpawnerScripts/EnemySpawner.cs
@@ -22,7 +22,7 @@ public class EnemySpawner : MonoBehaviour
 
     private EnemyFactory Enemy_Factory1;
 
-    public Transform[] SpawnPoints = null;
+    private Transform[] SpawnPoints = null;
 
     private Transform SpawnPoint;
 
@@ -73,16 +73,14 @@ public class EnemySpawner : MonoBehaviour
                 if (i > 10)
                 {
                     Enemy_Factory1.setSpawnPoint(SpawnPoint);
-                    e = (Enemy) Enemy_Factory1.produce();                        //180
-                    e.transform.rotation = Quaternion.Euler(new Vector3(180, 90, 0));
-                    //e.setSpeed(2);
+                    e = (Enemy) Enemy_Factory1.produce();                        
+                    e.transform.rotation = Quaternion.Euler(new Vector3(180, 90, 180));
                 }
                 else
                 {
                     Enemy_Factory.setSpawnPoint(SpawnPoint);
                     e = (Enemy) Enemy_Factory.produce();
-                    e.transform.rotation = Quaternion.Euler(new Vector3(0, 180, 2));
-                    //e.setSpeed(1);
+                    e.transform.rotation = Quaternion.Euler(new Vector3(0, 180, 0));
                 }
             }
