+++ ProjectSettings/TagManager.asset
@@ -16,7 +16,7 @@ TagManager:
   - Water
   - UI
   - Graphics
-  - 
+  - Enemy
   - 
   - 
   -
